Updated encoder implementation for AVR platform
The encoder switch would double count switch selects when large current draws occured as a result.  The PCB (v1.32) has a design flaw whereby the hardware debounce circuit is incorrectly designed resulting in partical hardware debouncing of the encoder switch.  This update addressed that by activating the (underpowered) internal pullups on the ATMEGA2560 as well as adding a debounce delay/check in software.